requirements.txt
gradio
torch
torchvision
mlflow
numpy
Pillow
scikit-image